+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## v0.1.8
 
 - `enumerate` の範囲指定のバグを修正
-- Add pathlib::PyPath
+- Add pathlib::PyPath::mkdir
 - Rename `Enumerator` -> `Iterable`
 
 ## v0.1.7